+++ src/bthdg/exch/Bitstamp.java
@@ -206,8 +206,8 @@ public class Bitstamp extends BaseExch {
         List<TradeData> trades = new ArrayList<TradeData>(len);
         for (int i = 0; i < len; i++) {
             JSONObject tObj = (JSONObject) array.get(i); // {"amount":"0.02500000","price":"683.00","tid":3406676,"date":"1391901009"}
-            double amount = Utils.getDouble(tObj.get("amount"));
-            double price = Utils.getDouble(tObj.get("price"));
+            float amount = Utils.getFloat(tObj.get("amount"));
+            float price = Utils.getFloat(tObj.get("price"));
             long timestamp = Utils.getLong(tObj.get("date")) * 1000;
             long tid = (Long) tObj.get("tid");
             TradeData tdata = new TradeData(amount, price, timestamp, tid, null);
