+++ src/bthdg/exch/TradeDataLight.java
@@ -3,10 +3,10 @@ package bthdg.exch;
 import bthdg.util.Utils;
 
 public class TradeDataLight {
-    public final double m_price;
+    public final float m_price;
     public final long m_timestamp;
 
-    public TradeDataLight(long timestamp, double price) {
+    public TradeDataLight(long timestamp, float price) {
         m_timestamp = timestamp;
         m_price = price;
     }
